Refactor: DocumentPanel-Sub-Komponenten in eigene Dateien extrahiert
EditorActionsInline (~70 Zeilen) und TransposeButtons (~25 Zeilen) waren
als innere Helper-Funktionen in DocumentPanel.tsx deklariert. Beide sind
strukturell eigenstaendig (EditorActionsInline liest ausschliesslich
useEditorCommands-State, TransposeButtons ist rein stateless) und
werden jetzt als separate Dateien gepflegt.

DocumentPanel von 650 auf 551 Zeilen.

Structural hygiene, keine Verhaltensaenderung und kein Profiling im
Hintergrund — der tatsaechliche Re-Render-Footprint bleibt gleich, weil
beide Komponenten im selben Render-Tree haengen.
